refactor(DeploymentViewer): break out components, refactor custom file toolbar
